Refactor #175 : user 쿼리 구조 리팩터링
- user 쿼리 구조 리팩터링
- 각 쿼리를 파일로 분리 및 index.ts에서 통합
