+++ client/src/queries/user/getDriverLocation.queries.ts
@@ -1,0 +1,16 @@
+import { gql } from '@apollo/client';
+
+const GET_DRIVER_LOCATION = gql`
+  query getDriverLocation($orderId: String!) {
+    getDriverLocation(orderId: $orderId) {
+      result
+      driverLocation {
+        lat
+        lng
+      }
+      error
+    }
+  }
+`;
+
+export default GET_DRIVER_LOCATION;
