+++ client/src/queries/user/getUserWithOrder.queries.ts
@@ -1,0 +1,28 @@
+import { gql } from '@apollo/client';
+
+const GET_USER_WITH_ORDER = gql`
+  query GetUserWithOrder {
+    getUserWithOrder {
+      result
+      user {
+        _id
+        email
+        type
+      }
+      order {
+        _id
+        startingPoint {
+          address
+          coordinates
+        }
+        destination {
+          address
+          coordinates
+        }
+      }
+      error
+    }
+  }
+`;
+
+export default GET_USER_WITH_ORDER;
