+++ client/src/routes/Home/Home.tsx
@@ -2,7 +2,7 @@ import React, { FC } from 'react';
 import { useCustomQuery } from '@hooks/useApollo';
 import { useHistory } from 'react-router-dom';
 
-import { GET_USER_WITH_ORDER } from '@queries/user.queries';
+import { GET_USER_WITH_ORDER } from '@queries/user';
 import { GetUserWithOrder } from '@/types/api';
 
 const Home: FC = () => {
